invoice_reader/requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
python-dateutil>=2.8.2